# Speed up writing records