Calling functions is costly in python, so duplicate some short code instead of calling it.

```
$ time scripts/extract-paired-reads.py ecoli_ref-5m.fastq -s a.se -p a.pe
67.88s user 28.85s system 93% cpu 1:43.77 total
$ git checkout master
$ time scripts/extract-paired-reads.py ecoli_ref-5m.fastq -s a.se -p a.pe
78.75s user 30.85s system 93% cpu 1:57.20 total
```

Happy 🦃 day!

- [x] Is it mergeable?
- [x] `make test` Did it pass the tests?
- [x] `make clean diff-cover` If it introduces new functionality in
  `scripts/` is it tested?
- [x] `make format diff_pylint_report cppcheck doc pydocstyle` Is it well
  formatted?
- [x] Did it change the command-line interface? Only backwards-compatible
  additions are allowed without a major version increment. Changing file
  formats also requires a major version number increment.
- [x] For substantial changes or changes to the command-line interface, is it
  documented in `CHANGELOG.md`? See [keepachangelog](http://keepachangelog.com/)
  for more details.
- [x] Was a spellchecker run on the source code and documentation after
  changes were made?
- [x] Do the changes respect streaming IO? (Are they
  tested for streaming IO?)