@betatim 2016-11-27
I went back to `%s` because "old habits". There was no benchmarking involved in the choice.