@betatim 2016-11-27
```python
rec_pair = '@%s\n%s\n+\n%s\n' * 2
a = ("asdasfsfasfasf", "fsgdfdfhdfhdfhdfhdfhdfhdfhdfergerbernrnten", "reerberbwefadsvsbrebsver")
%timeit rec_pair % (*a, *a)
#The slowest run took 6.07 times longer than the fastest. This could mean that an intermediate 
#result is being cached.
#1000000 loops, best of 3: 892 ns per loop
aa = (*a, *a)
%timeit rec_pair % aa
#The slowest run took 5.93 times longer than the fastest. This could mean that an intermediate 
#result is being cached.
#1000000 loops, best of 3: 643 ns per loop
rec_pair_fmt = '@{name1}\n{seq1}\n+\n{qual1}\n@{name2}\n{seq2}\n+\n{qual2}\n'
aa_fmt = {'name1': aa[0], 'qual1': aa[2], 'seq1': aa[1], 'name2': aa[0], 'qual2': aa[2], 'seq2': aa[1],}
%timeit rec_pair_fmt.format(**aa_fmt)
#1000000 loops, best of 3: 1.85 µs per loop
%timeit rec_pair_fmt.format(name1=aa[0], qual1=aa[2], seq1=aa[1], name2=aa[0], qual2=aa[2], seq2=aa[1])
#100000 loops, best of 3: 2.72 µs per loop
```

not sure I really believe what this is telling me. The recommended-by-the-BDFL way is slowing things down? Or is the benchmark cheating somehow? #grainofsalt